refactor: recoil 적용 및 App/Home 분리, isOpen 관련 로직 병합
